+++ components/credentials/credentials.js
@@ -1,0 +1,28 @@
+/*!
+ * Credentials module.
+ *
+ * Copyright (c) 2012-2015 Digital Bazaar, Inc. All rights reserved.
+ *
+ * @author Dave Longley
+ * @author David I. Lehn
+ */
+define([
+  'angular',
+  './credentials-routes'
+], function(angular, routes) {
+
+'use strict';
+
+var module = angular.module(
+  'bedrock.credentials', ['bedrock-credential-curator']);
+
+/* @ngInject */
+module.config(function($routeProvider) {
+  angular.forEach(routes, function(route) {
+    $routeProvider.when(route.path, route.options);
+  });
+});
+
+return module.name;
+
+});