+++ bedrock/pages/index.js
@@ -1,7 +1,8 @@
-var pages = {
-  dashboard: require('./dashboard'),
-  join: require('./join'),
-  navbar: require('./navbar'),
-  settings: require('./settings')
-};
+var pages = GLOBAL.bedrock.pages || {};
+
+pages.dashboard = require('./dashboard');
+pages.join = require('./join');
+pages.navbar = require('./navbar');
+pages.settings = require('./settings');
+
 module.exports = GLOBAL.bedrock.pages = pages;